server/pyproject.toml
[tool.poetry]
name = "receipt-recognition-system"
version = "0.1.0"
description = ""
authors = ["tyommmm <artyomhovsepyan135@gmail.com>"]
readme = "README.md"
packages = [{include = "*", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.6.2"
pytesseract = "0.3.10"
tesseract = "^0.1.3"
imutils = "^0.5.4"
opencv-python-headless = "^4.9.0.80"
scipy = "^1.12.0"
matplotlib = "^3.8.3"
scikit-image = "^0.22.0"
flask = "^3.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
